refactor: put tests in folders
