+++ Adaptors/MongoDB/tests/ResultWatcherTests.cs
@@ -17,7 +17,7 @@
 
 using ArmoniK.Core.Common.Storage;
 using ArmoniK.Core.Common.Storage.Events;
-using ArmoniK.Core.Common.Tests;
+using ArmoniK.Core.Common.Tests.TestBase;
 
 using Microsoft.Extensions.DependencyInjection;
