+++ Adaptors/S3/tests/ObjectStorageTests.cs
@@ -16,7 +16,7 @@
 // along with this program.  If not, see <http://www.gnu.org/licenses/>.
 
 using ArmoniK.Core.Common.Storage;
-using ArmoniK.Core.Common.Tests;
+using ArmoniK.Core.Common.Tests.TestBase;
 using ArmoniK.Core.Common.Utils;
 
 using Microsoft.Extensions.Configuration;
