+++ Common/tests/TestBase/AuthenticationTableTestBase.cs
@@ -29,7 +29,7 @@ using Microsoft.Extensions.Diagnostics.HealthChecks;
 
 using NUnit.Framework;
 
-namespace ArmoniK.Core.Common.Tests.Auth;
+namespace ArmoniK.Core.Common.Tests.TestBase;
 
 [TestFixture]
 public class AuthenticationTableTestBase
