+++ Common/tests/TestBase/ResultTableTestBase.cs
@@ -31,7 +31,7 @@ using Microsoft.Extensions.Diagnostics.HealthChecks;
 
 using NUnit.Framework;
 
-namespace ArmoniK.Core.Common.Tests;
+namespace ArmoniK.Core.Common.Tests.TestBase;
 
 [TestFixture]
 public class ResultTableTestBase
