+++ Common/tests/Validators/CancelTasksRequestValidatorTest.cs
@@ -20,7 +20,7 @@ using ArmoniK.Core.Common.gRPC.Validators;
 
 using NUnit.Framework;
 
-namespace ArmoniK.Core.Common.Tests;
+namespace ArmoniK.Core.Common.Tests.Validators;
 
 [TestFixture(TestOf = typeof(CancelTasksRequestValidator))]
 public class CancelTasksRequestValidatorTest
