+++ Common/tests/Validators/CreateLargeTaskRequestValidatorTest.cs
@@ -26,7 +26,7 @@ using Google.Protobuf.WellKnownTypes;
 
 using NUnit.Framework;
 
-namespace ArmoniK.Core.Common.Tests;
+namespace ArmoniK.Core.Common.Tests.Validators;
 
 [TestFixture(TestOf = typeof(CreateLargeTaskRequestValidator))]
 public class CreateLargeTaskRequestValidatorTest
