+++ Common/tests/Validators/CreateSmallTaskRequestValidatorTest.cs
@@ -27,7 +27,7 @@ using Google.Protobuf.WellKnownTypes;
 
 using NUnit.Framework;
 
-namespace ArmoniK.Core.Common.Tests;
+namespace ArmoniK.Core.Common.Tests.Validators;
 
 [TestFixture(TestOf = typeof(CreateSmallTaskRequestValidator))]
 public class CreateSmallTaskRequestValidatorTest
