+++ Common/tests/Validators/ListSessionsRequestValidatorTest.cs
@@ -22,7 +22,7 @@ using ArmoniK.Core.Common.gRPC.Validators;
 
 using NUnit.Framework;
 
-namespace ArmoniK.Core.Common.Tests;
+namespace ArmoniK.Core.Common.Tests.Validators;
 
 [TestFixture(TestOf = typeof(ListSessionsRequestValidator))]
 public class ListSessionsRequestValidatorTest
