+++ Common/tests/Validators/ListTasksRequestValidatorTest.cs
@@ -22,7 +22,7 @@ using ArmoniK.Core.Common.gRPC.Validators;
 
 using NUnit.Framework;
 
-namespace ArmoniK.Core.Common.Tests;
+namespace ArmoniK.Core.Common.Tests.Validators;
 
 [TestFixture(TestOf = typeof(ListTasksRequestValidator))]
 public class ListTasksRequestValidatorTest
