+++ Common/tests/Validators/SessionFilterValidatorTest.cs
@@ -21,7 +21,7 @@ using ArmoniK.Core.Common.gRPC.Validators;
 
 using NUnit.Framework;
 
-namespace ArmoniK.Core.Common.Tests;
+namespace ArmoniK.Core.Common.Tests.Validators;
 
 [TestFixture(TestOf = typeof(SessionFilterValidator))]
 public class SessionFilterValidatorTest
