+++ Common/tests/Validators/TaskFilterValidatorTest.cs
@@ -23,7 +23,7 @@ using ArmoniK.Core.Common.gRPC.Validators;
 
 using NUnit.Framework;
 
-namespace ArmoniK.Core.Common.Tests;
+namespace ArmoniK.Core.Common.Tests.Validators;
 
 [TestFixture(TestOf = typeof(TaskFilterValidator))]
 public class TaskFilterValidatorTest
