+++ Common/tests/Validators/TaskOptionsValidatorTest.cs
@@ -24,7 +24,7 @@ using Google.Protobuf.WellKnownTypes;
 
 using NUnit.Framework;
 
-namespace ArmoniK.Core.Common.Tests;
+namespace ArmoniK.Core.Common.Tests.Validators;
 
 [TestFixture(TestOf = typeof(TaskOptionsValidator))]
 public class TaskOptionsValidatorTest
